+++ client/src/app/details/(status)/page.tsx
